requirements.txt
boto3
python-dotenv
awscli
psycopg2
prettytable